Correcting data-types

+++ src/HRIS/EFCore/Entities/Position.cs
@@ -2,10 +2,10 @@ namespace Nova.HRIS.Entities;
 
 public class Position
 {
-    public short Id { get; set; }
+    public int Id { get; set; }
     public string Name { get; set; }
     public short Level { get; set; }
-    public short? ParentId { get; set; }
+    public int? ParentId { get; set; }
     public bool IsDeleted { get; set; }
     public int? InsertedById { get; set; }
     public DateTimeOffset? InsertedOn { get; set; }
